+++ Carlos/Errors.swift
@@ -10,6 +10,9 @@ public enum FetchError: ErrorType {
   /// Used when the specified key was invalid
   case InvalidKey
   
+  /// Used when some cached data was found but was likely corrupted
+  case InvalidCachedData
+  
   /// Used when the key doesn't satisfy the cache condition
   case ConditionNotSatisfied