+++ Carlos/NSKeyedUnarchiver+SwiftUtilities.m
@@ -22,4 +22,16 @@
   return object;
 }
 
++ (id) su_unarchiveObjectWithData:(NSData *)data {
+  id object = nil;
+  
+  @try {
+    object = [self unarchiveObjectWithData:data];
+  } @catch (NSException *exception) {
+    object = nil;
+  }
+  
+  return object;
+}
+
 @end