+++ Sample/ExamplesListViewController.swift
@@ -18,6 +18,7 @@ class ExamplesListViewController: UIViewController {
       Example(name: "Image cache", shortDescription: "Out-of-the-box image cache", segueIdentifier: "imageCache"),
       Example(name: "Data cache", shortDescription: "Out-of-the-box data cache", segueIdentifier: "dataCache"),
       Example(name: "JSON cache", shortDescription: "Out-of-the-box JSON cache", segueIdentifier: "jsonCache"),
+      Example(name: "User defaults cache", shortDescription: "Out-of-the-box NSUserDefaults cache", segueIdentifier: "userDefaultsCache"),
       Example(name: "Memory warnings", shortDescription: "Simple stack with memory warnings listeners", segueIdentifier: "memoryWarning")
     ]),
     ExamplesListSection(name: "Advanced", samples: [